+++ lib/service/FirebaseMessaging.dart
@@ -48,6 +48,7 @@ class FirebaseMessaging with Service implements NotificationsListener {
   static const String notifyConfigUpdate          = "edu.illinois.rokwire.firebase.messaging.config.update";
   static const String notifyPollOpen              = "edu.illinois.rokwire.firebase.messaging.poll.create";
   static const String notifyEventDetail           = "edu.illinois.rokwire.firebase.messaging.event.detail";
+  static const String notifyGameDetail            = "edu.illinois.rokwire.firebase.messaging.game.detail";
   static const String notifyAthleticsGameStarted  = "edu.illinois.rokwire.firebase.messaging.athletics_game.started";
   static const String notifySettingUpdated        = "edu.illinois.rokwire.firebase.messaging.setting.updated";
 
@@ -300,6 +301,9 @@ class FirebaseMessaging with Service implements NotificationsListener {
     }
     else if (type == "event_detail") {
       NotificationService().notify(notifyEventDetail, data);
+    }
+    else if (type == "game_detail") {
+      NotificationService().notify(notifyGameDetail, data);
     }
     else if (type == "athletics_game_started") {
       NotificationService().notify(notifyAthleticsGameStarted, data);